Convert to new override syntax
The override syntax has changed on the master branch of bitbake,
openembedded-core & poky. All other layers must be updated to use the
new syntax in order to build against the master branch.

Support for the new syntax has been backported to the bitbake branches
corresponding to the dunfell, gatesgarth and hardknott Yocto Project
releases. Support has not been added to the bitbake branch corresponding
to the zeus Yocto Project release so we need to drop zeus from
LAYERSERIES_COMPAT.

Signed-off-by: Paul Barker <paul.barker@sancloud.com>