feat(8723du): add driver RTL8723DU v5.13.4
Although there is a driver for RTL8723DU in newer kernels (>=6.2),
this kernel is not available with current Yocto releases. This kernel
version has not been released yet, either. Hence, it is useful to
have a driver for this chipset at hand for current kernels.

see: https://wireless.wiki.kernel.org/en/users/drivers/rtl819x

The branch v5.13.4 of the target repository does not compile with
Yocto `kirkstone`, so the master branch is used instead.